EW-627 added errors messages in the uc of personkontext

+++ src/modules/person/api/personenkontext.uc.ts
@@ -18,6 +18,7 @@ import { DomainError } from '../../../shared/error/domain.error.js';
 import { SchulConnexError } from '../../../shared/error/schul-connex.error.js';
 import { SchulConnexErrorMapper } from '../../../shared/error/schul-connex-error.mapper.js';
 import { DeletePersonenkontextDto } from './delete-personkontext.dto.js';
+import { MismatchedRevisionError } from '../../../shared/error/mismatched-revision.error.js';
 
 @Injectable()
 export class PersonenkontextUc {
@@ -133,6 +134,23 @@ export class PersonenkontextUc {
     public async deletePersonenkontextById(
         deletePersonenkontextDto: DeletePersonenkontextDto,
     ): Promise<void | SchulConnexError> {
+        const personkontext: Result<
+            PersonenkontextDo<true>,
+            DomainError
+        > = await this.personenkontextService.findPersonenkontextById(deletePersonenkontextDto.id);
+
+        if (!personkontext.ok) {
+            return SchulConnexErrorMapper.mapDomainErrorToSchulConnexError(personkontext.error);
+        }
+
+        if (personkontext.value.revision !== deletePersonenkontextDto.revision) {
+            return SchulConnexErrorMapper.mapDomainErrorToSchulConnexError(
+                new MismatchedRevisionError(
+                    `Revision ${deletePersonenkontextDto.revision} does not match revision ${personkontext.value.revision} of stored personenkontext.`,
+                ),
+            );
+        }
+
         const result: Result<void, DomainError> = await this.personenkontextService.deletePersonenkontextById(
             deletePersonenkontextDto.id,
         );
